Fix a leak where FileWatchers are added but never removed upon being expired from the cache, and some minor updates to privatize the cache by calling the public API in consumers
Prior to this change, FileWatchers to track when a KNN index is deleted are added to the ResourceWatcherService but never
removed even after indexes have been deleted, which causes the set of watchers to grow unbounded over time. This change
updates the logic to clean up FileWatchers upon cache eviction.